// apps/oidc/src/lib/common/openapi/oidc/models/AuthorizeRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * os-oidc
 * Aicacia OS OIDC API.
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ResponseMode } from './ResponseMode';
import {
	ResponseModeFromJSON,
	ResponseModeFromJSONTyped,
	ResponseModeToJSON,
	ResponseModeToJSONTyped
} from './ResponseMode';
import type { ResponseType } from './ResponseType';
import {
	ResponseTypeFromJSON,
	ResponseTypeFromJSONTyped,
	ResponseTypeToJSON,
	ResponseTypeToJSONTyped
} from './ResponseType';

/**
 *
 * @export
 * @interface AuthorizeRequest
 */
export interface AuthorizeRequest {
	/**
	 *
	 * @type {string}
	 * @memberof AuthorizeRequest
	 */
	clientId: string;
	/**
	 *
	 * @type {string}
	 * @memberof AuthorizeRequest
	 */
	codeChallenge?: string | null;
	/**
	 *
	 * @type {string}
	 * @memberof AuthorizeRequest
	 */
	codeChallengeMethod?: string | null;
	/**
	 *
	 * @type {string}
	 * @memberof AuthorizeRequest
	 */
	nonce?: string | null;
	/**
	 *
	 * @type {string}
	 * @memberof AuthorizeRequest
	 */
	redirectUri: string;
	/**
	 *
	 * @type {string}
	 * @memberof AuthorizeRequest
	 */
	registration?: string | null;
	/**
	 *
	 * @type {ResponseMode}
	 * @memberof AuthorizeRequest
	 */
	responseMode: ResponseMode;
	/**
	 *
	 * @type {ResponseType}
	 * @memberof AuthorizeRequest
	 */
	responseType: ResponseType;
	/**
	 *
	 * @type {string}
	 * @memberof AuthorizeRequest
	 */
	scope: string;
	/**
	 *
	 * @type {string}
	 * @memberof AuthorizeRequest
	 */
	state?: string | null;
}

/**
 * Check if a given object implements the AuthorizeRequest interface.
 */
export function instanceOfAuthorizeRequest(value: object): value is AuthorizeRequest {
	if (!('clientId' in value) || value['clientId'] === undefined) return false;
	if (!('redirectUri' in value) || value['redirectUri'] === undefined) return false;
	if (!('responseMode' in value) || value['responseMode'] === undefined) return false;
	if (!('responseType' in value) || value['responseType'] === undefined) return false;
	if (!('scope' in value) || value['scope'] === undefined) return false;
	return true;
}

export function AuthorizeRequestFromJSON(json: any): AuthorizeRequest {
	return AuthorizeRequestFromJSONTyped(json, false);
}

export function AuthorizeRequestFromJSONTyped(
	json: any,
	ignoreDiscriminator: boolean
): AuthorizeRequest {
	if (json == null) {
		return json;
	}
	return {
		clientId: json['client_id'],
		codeChallenge: json['code_challenge'] == null ? undefined : json['code_challenge'],
		codeChallengeMethod:
			json['code_challenge_method'] == null ? undefined : json['code_challenge_method'],
		nonce: json['nonce'] == null ? undefined : json['nonce'],
		redirectUri: json['redirect_uri'],
		registration: json['registration'] == null ? undefined : json['registration'],
		responseMode: ResponseModeFromJSON(json['response_mode']),
		responseType: ResponseTypeFromJSON(json['response_type']),
		scope: json['scope'],
		state: json['state'] == null ? undefined : json['state']
	};
}

export function AuthorizeRequestToJSON(json: any): AuthorizeRequest {
	return AuthorizeRequestToJSONTyped(json, false);
}

export function AuthorizeRequestToJSONTyped(
	value?: AuthorizeRequest | null,
	ignoreDiscriminator: boolean = false
): any {
	if (value == null) {
		return value;
	}

	return {
		client_id: value['clientId'],
		code_challenge: value['codeChallenge'],
		code_challenge_method: value['codeChallengeMethod'],
		nonce: value['nonce'],
		redirect_uri: value['redirectUri'],
		registration: value['registration'],
		response_mode: ResponseModeToJSON(value['responseMode']),
		response_type: ResponseTypeToJSON(value['responseType']),
		scope: value['scope'],
		state: value['state']
	};
}